% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{choose_knots_even}
\alias{choose_knots_even}
\title{choose_knots_even}
\usage{
choose_knots_even(X, num_knots)
}
\arguments{
\item{num_knots}{the number of knots to use}
}
\value{
a vector containing the locations of the knots
}
\description{
A function to place equally spaced knots for fitting b-splines
}
