% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{choose_knots_quantile}
\alias{choose_knots_quantile}
\title{choose_knots_quantile}
\usage{
choose_knots_quantile(x, num_knots)
}
\arguments{
\item{x}{vector of treatment doses}

\item{num_knots}{the number of knots to use}
}
\value{
a vector containing the locations of the knots
}
\description{
A function to choose knots for fitting b-splines by the quantile of x
}
