% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcont_did.R
\name{ggcont_did}
\alias{ggcont_did}
\title{Plot Results with a Continuous Treatment}
\usage{
ggcont_did(dose_obj, type = "att")
}
\arguments{
\item{dose_obj}{a result from running \code{cont_did}}

\item{type}{whether to plot ATT(d) or ACRT(d), defaults to \code{att} for
plotting ATT(d).  For ACRT(d), use "acrt"}
}
\description{
a function to plot results with a continuous treatment
}
