% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npiv_regression.R
\name{npiv_regression}
\alias{npiv_regression}
\title{Non-parametric Instrumental Variable Regression}
\usage{
npiv_regression(
  data,
  treatment_col,
  outcome_col,
  id_col = NULL,
  alpha = c(0.05),
  nx = 1000,
  nL = 9,
  r = 4,
  M = 5
)
}
\arguments{
\item{data}{A data frame containing the variables for the regression.}

\item{treatment_col}{Name of the treatment variable column.}

\item{outcome_col}{Name of the outcome variable column.}

\item{id_col}{Name of the ID variable column (optional).}

\item{alpha}{Significance level(s) for confidence intervals.}

\item{nx}{Number of points for the grid of x values.}

\item{nL}{Maximum resolution level for J.}

\item{r}{B-spline order.}

\item{M}{Ex ante upper bound on sup_x h_0(x).}
}
\value{
A list containing the results of the NPIV regression.
}
\description{
Non-parametric Instrumental Variable Regression
}
