% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_pte_cont.R
\name{setup_pte_cont}
\alias{setup_pte_cont}
\title{setup_pte_cont}
\usage{
setup_pte_cont(
  yname,
  gname,
  tname,
  idname,
  data,
  xformula = ~1,
  target_parameter,
  aggregation,
  treatment_type,
  required_pre_periods = 1,
  anticipation = 0,
  base_period = "varying",
  cband = TRUE,
  alp = 0.05,
  boot_type = "multiplier",
  weightsname = NULL,
  gt_type = "att",
  ret_quantile = 0.5,
  biters = 100,
  cl = 1,
  dname,
  dvals = NULL,
  degree = 1,
  num_knots = 0,
  ...
)
}
\arguments{
\item{dvals}{an optional argument specifying which values of the
treatment to evaluate ATT(d) and/or ACRT(d).  If no values are
supplied, then the default behavior is to set
\code{dvals} to be the 1st to 99th percentiles of the dose among
units that experience any positive dose.}
}
\value{
\code{pte_params} object
}
\description{
A function that creates a pte_params object, adding
several different variables that are needed when there is a
continuous treatment.
}
