% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_package}
\alias{update_package}
\title{Update and Reinstall Package}
\usage{
update_package(pkg = "contdid", restart = FALSE)
}
\arguments{
\item{pkg}{The name of the package}

\item{restart}{Whether to restart the R session after reinstalling (default: FALSE)}
}
\description{
This function automates the process of updating and reinstalling the package during development.
}
