% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{contdid-package}
\alias{contdid}
\alias{contdid-package}
\title{Difference-in-differences with a continuous treatment}
\description{
\code{contdid}  is package for estimating the effect of a continuous treatment in a difference-in-differences framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bcallaway11.github.io/contdid/}
}

}
\author{
\strong{Maintainer}: Brantly Callaway \email{brantly.callaway@uga.edu}

}
\keyword{internal}
